support/package-builder: add the ability to build Photon-3 in Photon-4
Use Photon-3 docker image while installing packages in chroot if host
rpm doesn't support bdb.

Makefile:
- Use tabs instead of spaces all over
- Add support for importing photon3 docker image if neeeded as a
  pre-requisite for build
- Remove PHOTON_DOCKER_IMAGE env variable support
- Use --update option with createrepo, reduces I/O operations
  significantly
- Move docker related checks to check-tools target

- tools/src:
- Use makefile to build contain_priv

support/dockerfiles:
- Remove unused ubuntu-dev directory
- Use approproate FROMs in docker files

tools/scripts:
- Add helper script to handle importing Photon-3 docker image and
  building photon_builder container image

support/image-builder/ostree-tools:
- Refactor the script to use rpms from local repo
- Create ostree docker image from fresh everytime

support/image-builder/iso:
- Fix mk-install-iso.sh to handle bdb dependency properly.

support/make/makedefs.mk
- Add options related to docker images
- Add --update option to createrepo option

Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/17096
Tested-by: gerrit-photon <photon-checkins@vmware.com>

+++ support/image-builder/iso/mk-install-iso.sh
@@ -1,18 +1,6 @@
 #! /bin/bash
-#################################################
-#       Title:  mk-install-iso                  #
-#        Date:  2014-11-26                      #
-#     Version:  1.0                             #
-#      Author:  dthaluru@vmware.com             #
-#     Options:                                  #
-#################################################
-#   Overview
-#       Generates a photon iso
-#   End
-#
 
-set -e
-set -x
+set -ex
 
 SCRIPT_PATH=$(dirname $(realpath -s $0))
 PRGNAME=${0##*/}    # script name minus the path
@@ -60,12 +48,21 @@ clean_requirements_on_remove=true
 repodir=${WORKINGDIR}
 EOF
 
-rpm --root $INITRD --initdb --dbpath /var/lib/rpm
+rpmdb_init_cmd="rpm --root ${INITRD} --initdb --dbpath /var/lib/rpm"
+# PHOTON_BUILDER_TAG is exported from top level Makefile
+if [ "$(rpm -E %{_db_backend})" != "bdb" ]; then
+  rpmdb_init_cmd="docker run --rm -v ${INITRD}:${INITRD} ${PHOTON_BUILDER_TAG} /bin/bash -c \"${rpmdb_init_cmd}\""
+fi
 
-TDNF_CMD="tdnf install -y --releasever ${PHOTON_RELEASE_VERSION} --installroot $INITRD --rpmverbosity 10 -c ${WORKINGDIR}/tdnf.conf -q $PACKAGES"
+if ! eval "${rpmdb_init_cmd}"; then
+  echo "ERROR: failed to initialize rpmdb" 1>&2
+  exit 1
+fi
+
+TDNF_CMD="tdnf install -y --releasever ${PHOTON_RELEASE_VERSION} --installroot $INITRD --rpmverbosity error -c ${WORKINGDIR}/tdnf.conf $PACKAGES"
 
 # run host's tdnf, if fails - try one from photon:latest docker image
-$TDNF_CMD || docker run -v $RPMS_PATH:$RPMS_PATH -v $WORKINGDIR:$WORKINGDIR $PHOTON_DOCKER_IMAGE $TDNF_CMD
+$TDNF_CMD || docker run --rm -v $RPMS_PATH:$RPMS_PATH -v $WORKINGDIR:$WORKINGDIR $PHOTON_DOCKER_IMAGE $TDNF_CMD
 
 rm -f ${WORKINGDIR}/photon-local.repo ${WORKINGDIR}/tdnf.conf
